added light power switching and busser alarm

+++ PowerController.h
@@ -12,7 +12,7 @@ public:
   void activateFans(float overTemp);
 private:
   // Set PWM as percent 0..100 (will be constrained and converted to 0..255)
-  void setPwm(uint8_t percent);
+  void setPwm(float percent);
 };
 
 #endif
